CRC module of stm32f0/3 renamed to crc_v2
