+++ include/libopencm3/stm32/f3/crc.h
@@ -32,6 +32,6 @@
 #ifndef LIBOPENCM3_CRC_H
 #define LIBOPENCM3_CRC_H
 
-#include <libopencm3/stm32/common/crc_common_f03.h>
+#include <libopencm3/stm32/common/crc_v2.h>
 
 #endif
