+++ lib/stm32/common/crc_v2.c
