+++ lib/stm32/f3/Makefile
@@ -39,7 +39,7 @@ ARFLAGS		= rcs
 OBJS		= rcc.o adc.o i2c.o usart.o dma.o flash.o
 
 OBJS            += gpio_common_all.o gpio_common_f0234.o \
-		   dac_common_all.o usart_common_all.o crc_common_all.o crc_common_f03.o \
+		   dac_common_all.o usart_common_all.o crc_common_all.o crc_v2.o \
 		   iwdg_common_all.o spi_common_all.o dma_common_l1f013.o\
 		   timer_common_all.o timer_common_f234.o flash_common_f234.o \
 		   flash.o exti_common_all.o rcc_common_all.o spi_common_f03.o
